Updated README.md

+++ main.cpp
@@ -1,10 +1,15 @@
 #include<iostream>
 #include<string>
 #include<cstdlib>
+#include<filesystem>
 
 void convertFile(const std::string& inputFile, const std::string& outputFile) {
+  if(!std::filesystem::exists(inputFile)) {
+    std::cerr << "Arquivo de entrada não encontrado." << std::endl;
+    return;
+  }
   std::string command = "ffmpeg -i \"" + inputFile + "\" \"" + outputFile + "\"";
-
+  
   int result = std::system(command.c_str());
   if(result == 0) {
     std::cout << "Conversão feita com sucesso: " << outputFile << std::endl;
